refactor: extract alc-pdf crate + achieve 100% coverage (#phase2)
- Move PDF generation (printpdf) into crates/alc-pdf (types + render)
- Route handlers remain in main crate, import generate_pdf from alc-pdf
- Shared types (RestraintReportResponse etc.) now defined in alc-pdf::types
- Remove printpdf dependency from root Cargo.toml
- Add 5 tests: overlap, weekly_subtotals, month>=4, list_drivers DB errors
- render.rs + dtako_restraint_report_pdf.rs both 100% → registered in coverage_100.toml (50 files)
