+++ crates/alc-pdf/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "alc-pdf"
+version = "0.1.0"
+edition = "2021"
+
+[lints.rust]
+unexpected_cfgs = { level = "allow", check-cfg = ['cfg(coverage)'] }
+
+[dependencies]
+chrono = { version = "0.4", features = ["serde"] }
+printpdf = "0.8"
+serde = { version = "1", features = ["derive"] }
+uuid = { version = "1", features = ["v4", "serde"] }
