+++ crates/alc-pdf/src/types.rs
@@ -1,0 +1,81 @@
+use chrono::NaiveDate;
+use serde::Serialize;
+use uuid::Uuid;
+
+#[derive(Debug, Serialize)]
+pub struct RestraintReportResponse {
+    pub driver_id: Uuid,
+    pub driver_name: String,
+    pub year: i32,
+    pub month: u32,
+    pub max_restraint_minutes: i32,
+    pub days: Vec<RestraintDayRow>,
+    pub weekly_subtotals: Vec<WeeklySubtotal>,
+    pub monthly_total: MonthlyTotal,
+}
+
+#[derive(Debug, Serialize)]
+pub struct RestraintDayRow {
+    pub date: NaiveDate,
+    pub is_holiday: bool,
+    pub start_time: Option<String>,
+    pub end_time: Option<String>,
+    pub operations: Vec<OperationDetail>,
+    pub drive_minutes: i32,
+    pub cargo_minutes: i32,
+    pub break_minutes: i32,
+    pub restraint_total_minutes: i32,
+    pub restraint_cumulative_minutes: i32,
+    pub drive_average_minutes: f64,
+    pub rest_period_minutes: Option<i32>,
+    pub remarks: String,
+    // CSV互換フィールド
+    pub overlap_drive_minutes: i32,
+    pub overlap_cargo_minutes: i32,
+    pub overlap_break_minutes: i32,
+    pub overlap_restraint_minutes: i32,
+    pub restraint_main_minutes: i32,
+    pub drive_avg_before: Option<i32>,
+    pub drive_avg_after: Option<i32>,
+    pub actual_work_minutes: i32,
+    pub overtime_minutes: i32,
+    pub late_night_minutes: i32,
+    pub overtime_late_night_minutes: i32,
+}
+
+#[derive(Debug, Clone, Serialize)]
+pub struct OperationDetail {
+    pub unko_no: String,
+    pub drive_minutes: i32,
+    pub cargo_minutes: i32,
+    pub break_minutes: i32,
+    pub restraint_minutes: i32,
+}
+
+#[derive(Debug, Serialize)]
+pub struct WeeklySubtotal {
+    pub week_end_date: NaiveDate,
+    pub drive_minutes: i32,
+    pub cargo_minutes: i32,
+    pub break_minutes: i32,
+    pub restraint_minutes: i32,
+}
+
+#[derive(Debug, Serialize)]
+pub struct MonthlyTotal {
+    pub drive_minutes: i32,
+    pub cargo_minutes: i32,
+    pub break_minutes: i32,
+    pub restraint_minutes: i32,
+    pub fiscal_year_cumulative_minutes: i32,
+    pub fiscal_year_total_minutes: i32,
+    // CSV互換フィールド
+    pub overlap_drive_minutes: i32,
+    pub overlap_cargo_minutes: i32,
+    pub overlap_break_minutes: i32,
+    pub overlap_restraint_minutes: i32,
+    pub actual_work_minutes: i32,
+    pub overtime_minutes: i32,
+    pub late_night_minutes: i32,
+    pub overtime_late_night_minutes: i32,
+}
